+++ src/vscode_code_action/ez_code_action.ts
@@ -13,7 +13,7 @@ export function registerEzAction(context: vscode.ExtensionContext) {
     providers.push(new StringConvertFixer())
     providers.push(new GraphqlToDartApiFixer())
 
-    // providers.push(new ClassQuickFix())
+    providers.push(new ClassQuickFix())
     // providers.push(new DartCurserDetector())
     for (let p of providers) {
         // 註冊命令回調